Update TypeSafeIndex semantics
1. Allow comparison with size_t values. This includes explicit white listing
   of unsigned values and binary versions of the comparisons.
2. Update the unit tests with more meaningful results; the old unit tests
   tested for the existence of functions. However, those functions proved
   to generate warnings which, in turn, generated errors.
3. Update comparison declarations to be *const*. This ended up breaking
   a single test in which indices of two different types are compared --
   something that the former unit tests suggested shouldn't have been
   possible.
4. Update the tests to produce more concrete and reliable results on the
   semantics.
5. Update the documentation underscoring C++'s implicit narrowing
   conversions that allow instantiation of TypeSafeIndex from non `int`
   integral types.

+++ multibody/multibody_tree/test/multibody_tree_test.cc
@@ -142,7 +142,8 @@ void VerifyModelBasics(const MultibodyTree<T>& model) {
     // We added actuators and joints in the same order. Assert this before
     // making that assumption in the test that follows.
     const Joint<T>& joint = actuator.joint();
-    ASSERT_EQ(actuator.index(), joint.index());
+    ASSERT_EQ(static_cast<int>(actuator.index()),
+              static_cast<int>(joint.index()));
     const std::string& joint_name = kJointNames[names_index];
     EXPECT_EQ(joint.name(), joint_name);
     ++names_index;
